requirements.txt
crewai
langgraph
langchain
langchain-openai
streamlit
pydantic
python-dotenv